Moved enums into subfolder
