+++ app/Enums/ClaimType.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App;
+namespace App\Enums;
 
 enum ClaimType: string
 {
