+++ app/Enums/TaskFrequency.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App;
+namespace App\Enums;
 
 enum TaskFrequency: string
 {
